cleanup code and remove debugs
