+++ src/main/resources/assets/skyblock-pv/lang/en_us.json
@@ -3,5 +3,6 @@
     "skyblockpv.tab.combat": "Combat",
     "skyblockpv.tab.collection": "Collections",
     "skyblockpv.tab.mining": "Mining",
-    "skyblockpv.tab.inventory": "Inventory"
+    "skyblockpv.tab.inventory": "Inventory",
+    "skyblockpv.tab.fishing": "Fishing"
 }
